+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Visualization Python API
   - Bbox
     (<https://github.com/openvinotoolkit/datumaro/pull/744>)
+- Support for exclusive of labels with LabelGroup
+  (<https://github.com/openvinotoolkit/datumaro/pull/742>)
 
 ### Changed
 - N/A